% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mcmodummies.R
\name{create_mcmodummies}
\alias{create_mcmodummies}
\title{Creates dummy variables for multiple choice, multiple option question}
\usage{
create_mcmodummies(item, ...)
}
\description{
This function creates dummy variables for a multiple choice, multiple option question
}
