% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matchvars.R
\name{get_matchvars}
\alias{get_matchvars}
\title{Gets list of variables with a certain substring in them}
\usage{
get_matchvars(d, pattern)
}
\description{
Searches through all variable names and returns all that match a search term
}
