% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmodummies.R
\name{get_mcmodummies}
\alias{get_mcmodummies}
\title{Creates MCMO dummy variables}
\usage{
get_mcmodummies(item, ...)
}
\arguments{
\item{item}{Data vector from MCMO question}

\item{...}{Other options as in \code{get_mcmoopts()}}
}
\description{
This function creates dummy variables for a multiple choice, multiple option
question. Suggested data check: Run \code{get_mcmoopts()} first to see what
survey choices are being found, if you are not using custom survey choices.
}
\examples{
# Print dummy variables of MCMO question "Q3" in "survey" data.frame object
get_mcmodummies(survey$Q3)
# Omit "Computer science" option
get_mcmodummies(survey$Q3, omit="Computer science")

}
