% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmoopts.R
\name{get_mcmoopts}
\alias{get_mcmoopts}
\title{Gets options for multiple choice, multiple option question}
\usage{
get_mcmoopts(item, delim = ",", omit = c())
}
\description{
This function gets all options for a multiple choice, multiple option question
}
