% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcmopct.R
\name{get_mcmopct}
\alias{get_mcmopct}
\title{Gets percentage from MCMO question}
\usage{
get_mcmopct(item, value, digits = 2)
}
\description{
Gets percentage of values in a multiple choice question with specific value
}
