% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcsopct.R
\name{get_mcsopct}
\alias{get_mcsopct}
\title{Gets percentage from MCSO question}
\usage{
get_mcsopct(item, value, digits = 2)
}
\description{
Gets percentage of values in a multiple choice question with specific value
}
