% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_textresp.R
\name{get_textresp}
\alias{get_textresp}
\title{Returns text responses to a text-based question}
\usage{
get_textresp(item, rm.nl = F, empty = NA)
}
\description{
This function returns, in a vector, all of the text responses from a text-based question
}
