% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_linkednum.R
\name{graph_linkednum}
\alias{graph_linkednum}
\title{Graphs multiple numeric questions}
\usage{
graph_linkednum(items, labels, sort = "entry", totalcheck = NA,
  central = "mean", spread = "sd1", fill = "tomato1")
}
\description{
This function creates bar plots with error bars for linked numeric questions
}
