% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_matrix.R
\name{graph_matrix}
\alias{graph_matrix}
\title{Graphs a multiple choice, single option question}
\usage{
graph_matrix(items, levels, labels, sort = "entry", palette = "ordered",
  ...)
}
\description{
This function creates a horizontal bar graph for a multiple choice, single option question
}
