% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_mcmo.R
\name{graph_mcmo}
\alias{graph_mcmo}
\title{Graphs a multiple choice, multiple option question}
\usage{
graph_mcmo(item, sort = "alpha", custom = NULL, stat = "count", ...)
}
\description{
This function creates a horizontal bar graph for a multiple choice, multiple option question
}
