% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_strings.R
\name{wrap_strings}
\alias{wrap_strings}
\title{A vector of strings wrapping function}
\usage{
wrap_strings(vector_of_strings, ...)
}
\description{
This function applies wrap_sentence to a vector of strings
}
