% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mcsopct.R
\name{get_mcsopct}
\alias{get_mcsopct}
\title{Gets percentage from MCSO question}
\usage{
get_mcsopct(item, value, digits = 2)
}
\arguments{
\item{item}{Data vector from MCSO question}

\item{value}{The particular value being matched}

\item{digits}{The number of digits after the decimal for rounding the
percent}
}
\description{
Gets percentage of values in a MCSO question with specific value. It is
more often used for data checking or reporting single numbers in a document.
}
\examples{
# Gets the percent of people who answered "Male" as an option
get_mcmopct(survey$Q4, "Male")
# Rounds to the nearest integer
get_mcmopct(survey$Q4, "Male", digits=0)

}
