% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numericpct.R
\name{get_numericpct}
\alias{get_numericpct}
\title{Gets percentages in a numeric variable}
\usage{
get_numericpct(item, min, max = min, digits = 2)
}
\description{
This function gets the percent of values in a range, inclusive
}
