% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_mcso.R
\name{graph_mcso}
\alias{graph_mcso}
\title{Graphs a multiple choice, single option question}
\usage{
graph_mcso(item, sort = "alpha")
}
\arguments{
\item{item}{The vector or column from the dataset with the options
graph_mcso()}
}
\description{
This function creates a horizontal bar graph for a multiple choice, single option question
}
