% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_numeric.R
\name{graph_numeric}
\alias{graph_numeric}
\title{Graphs a numeric question}
\usage{
graph_numeric(item, unit, xmin = 0, xmax = max(item) + binwidth,
  binwidth = 1)
}
\description{
This function creates a histogram for a numeric question
}
