% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_sentence.R
\name{wrap_sentence}
\alias{wrap_sentence}
\title{A string wrapping function}
\usage{
wrap_sentence(string, width = 30)
}
\description{
This function takes a space-delimited sentence and inserts newlines to wrap the sentence.
}
